rename favourite.rs to trusted.rs
