+++ discovery_engine_core/core/src/stack/ops/trusted.rs
